requirements.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
libsass==0.13.2
MarkupSafe==1.0
six==1.10.0
Werkzeug==0.12.2
